feat: improved architecture
