+++ schemas/Schemas.py
@@ -84,44 +84,6 @@ class ClicksSummaryByCountry(BaseModel):
         from_attributes = True
 
 
-class ShareTribeUserAttributesProfile(BaseModel):
-    displayName: str
-    abbreviatedName: str
-
-class ShareTribeUserAttributes(BaseModel):
-    email: str
-    profile: ShareTribeUserAttributesProfile
-
-class ShareTribeUser(BaseModel):
-    id: str
-    type: str
-    attributes: ShareTribeUserAttributes
-
-    class Config:
-        from_attributes = True
-
-class ShareTribeIncludedImageAttributesVariant(BaseModel):
-    url: str
-    width: int
-    height: int
-
-class ShareTribeIncludedImageAttributes(BaseModel):
-    variants: dict[str, ShareTribeIncludedImageAttributesVariant]
-
-class ShareTribeIncludedImage(BaseModel):
-    id: str
-    type: str
-    attributes: ShareTribeIncludedImageAttributes
-
-    class Config:
-        from_attributes = True
-
-class ShareTribeUserResponse(BaseModel):
-    data: ShareTribeUser
-    included: List[ShareTribeIncludedImage] | None = None
-
-    class Config:
-        from_attributes = True
 
 class IPInfo(BaseModel):
     id: int
@@ -145,3 +107,19 @@ class ShortenLinkResponse(BaseModel):
 
     class Config:
         from_attributes = True
+
+class UserBody(BaseModel):
+    username: str
+    password: str
+    
+    class Config:
+        form_attributes = True  
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+    
+class ShortenLinkRequest(BaseModel):
+    title: str
+    long_url: str
+    custom_back_half: Optional[str] = None
